- added shader compilation cache for the ShaderPermutator, as the D3D11 shader compilation took quite a bit a lot. OpenGL is broken now.

+++ libs/sge_utils/src/sge_utils/utils/hash_combine.h
@@ -9,7 +9,7 @@ inline unsigned int hashCString_djb2(const char* str) {
 
 	unsigned int hash = 5381;
 
-	int iOffset = 0;
+	size_t iOffset = 0;
 	while (str[iOffset] != '\0') {
 		hash = ((hash << 5) + hash) + (str[iOffset]); /* hash * 33 + c */
 		iOffset++;
@@ -18,13 +18,13 @@ inline unsigned int hashCString_djb2(const char* str) {
 	return hash;
 }
 
-inline unsigned int hash_djb2(const char* const mem, const int numBytes) {
+inline unsigned int hash_djb2(const char* const mem, const size_t numBytes) {
 	if (!mem) {
 		return 0;
 	}
 
 	unsigned int hash = 5381;
-	for (int iByte = 0; iByte < numBytes; ++iByte) {
+	for (size_t iByte = 0; iByte < numBytes; ++iByte) {
 		int c = int(mem[iByte]);
 		hash = ((hash << 5) + hash) + c; /* hash * 33 + c */
 	}
